refactor(main): extract update service runtime from main.ts
